+++ shell/imports/server/persistent.js
@@ -14,6 +14,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { inMeteor } from "/imports/server/async-helpers.js";
+
 const Crypto = Npm.require("crypto");
 
 const privateDb = Symbol("PersistentImpl.db");
