+++ shell/server/demo-server.js
@@ -14,6 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { waitPromise } from "/imports/server/async-helpers.js";
 import { allowDemo } from "/imports/demo.js";
 
 const DEMO_EXPIRATION_MS = 60 * 60 * 1000;
